% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{body_outputBulkRNA_countONly}
\alias{body_outputBulkRNA_countONly}
\title{body_outputBulkRNA_countONly}
\usage{
body_outputBulkRNA_countONly()
}
\value{

}
\description{
body_outputBulkRNA_countONly
}
