% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{count_transform}
\alias{count_transform}
\title{Title count_transform}
\usage{
count_transform(count_ns, pickGeneDFall, refDataExp, weight_table)
}
\arguments{
\item{weight_table}{}
}
\value{

}
\description{
Title count_transform
}
