% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{draw_BoxPlot}
\alias{draw_BoxPlot}
\title{draw_BoxPlot}
\usage{
draw_BoxPlot(
  obj_in,
  group.by = "diag",
  features = c("CRLF2"),
  cols = subtypeCol(),
  assay_name_in = "vst",
  plot_title = "Box Plot",
  x_lab = "Subtype",
  y_lab = NULL,
  useGeneName = T,
  df_geneName = info_gtf_hg38,
  highlightLevel = NULL,
  sizeHighlight = 1.5
)
}
\arguments{
\item{sizeHighlight}{}
}
\value{

}
\description{
draw_BoxPlot
}
