% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{draw_DimPlot}
\alias{draw_DimPlot}
\title{draw_DimPlot}
\usage{
draw_DimPlot(
  obj_in,
  group.by = "diag",
  cols = subtypeCol(),
  axis_by = 10,
  reduction = "tsne",
  plot_title = NULL,
  legend_title = "Subtypes",
  highlightLevel = NULL,
  sizeHighlight = 2
)
}
\arguments{
\item{plot_title}{}
}
\value{

}
\description{
draw_DimPlot
}
