% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilits.R
\name{f_imputation}
\alias{f_imputation}
\title{f_imputation}
\usage{
f_imputation(
  obj_ref,
  assay_name_in = "keyFeatures",
  df_in,
  method = "randomforest"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
f_imputation
}
