% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_RNAseqCNV.R
\name{get_arm_metr}
\alias{get_arm_metr}
\title{Title get_arm_metr}
\usage{
get_arm_metr(count_ns, smpSNPdata, sample_name, centr_ref)
}
\arguments{
\item{centr_ref}{}
}
\value{

}
\description{
Title get_arm_metr
}
