% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilits.R
\name{get_embeding_feature}
\alias{get_embeding_feature}
\title{get_embeding_feature}
\usage{
get_embeding_feature(
  obj_in,
  assay_name_in = "vst",
  features = "cluster_Phenograph_pca1",
  reduction = "tsne"
)
}
\arguments{
\item{reduction}{}
}
\value{

}
\description{
get_embeding_feature
}
