% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{get_GEP_pred}
\alias{get_GEP_pred}
\title{get_GEP_pred}
\usage{
get_GEP_pred(
  count_matrix,
  featureN_PG = c(100, 200, 300, 400, 500, 600, 700, 800, 900, 1000, 1058)
)
}
\arguments{
\item{featureN_PG}{}
}
\value{

}
\description{
get_GEP_pred
}
