% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{get_norm_exp}
\alias{get_norm_exp}
\title{get_norm_exp}
\usage{
get_norm_exp(
  count_table,
  standard_samples,
  minReadCnt,
  samp_prop,
  weight_table,
  weight_samp_prop,
  generate_weights
)
}
\arguments{
\item{generate_weights}{}
}
\value{

}
\description{
get_norm_exp
}
