% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{get_PhenoGraphPred}
\alias{get_PhenoGraphPred}
\title{get_PhenoGraphPred}
\usage{
get_PhenoGraphPred(
  obj_in,
  panelName = "PhenographPred",
  SampleLevel = "TestSample",
  type = "value"
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
get_PhenoGraphPred
}
