% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{get_PhenoGraphPreds}
\alias{get_PhenoGraphPreds}
\title{get_PhenoGraphPreds}
\usage{
get_PhenoGraphPreds(
  obj_in,
  feature_panel = "keyFeatures",
  SampleLevel = "TestSample",
  variable_n_list = c(seq(100, 1000, 100), 1058),
  neighbor_k = 10,
  cores = 1
)
}
\arguments{
\item{cores}{}
}
\value{

}
\description{
get_PhenoGraphPreds
}
