% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC.R
\name{get_SC_subtypes}
\alias{get_SC_subtypes}
\title{get_SC_subtypes}
\usage{
get_SC_subtypes(
  count_matrix,
  levelsIn = c("Pro-B1", "Pro-B2", "Pre-B1", "Pre-B2"),
  SE_celltype = SE_celltype,
  SE_BALL = SE_BALL
)
}
\arguments{
\item{SE_BALL}{}
}
\value{

}
\description{
get_SC_subtypes
}
