% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC.R
\name{get_singleR_heatmapScore}
\alias{get_singleR_heatmapScore}
\title{get_singleR_heatmapScore}
\usage{
get_singleR_heatmapScore(singleR_out)
}
\arguments{
\item{singleR_out}{}
}
\value{

}
\description{
get_singleR_heatmapScore
}
