% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC.R
\name{get_singleR_label}
\alias{get_singleR_label}
\title{get_singleR_label}
\usage{
get_singleR_label(
  singleR_out,
  obj_in = NULL,
  reduction_method = "umap",
  levelsIn = c("Pro-B1", "Pro-B2", "Pre-B1", "Pre-B2"),
  label_levelsIn = "Pro/Pre-B"
)
}
\arguments{
\item{label_levelsIn}{}
}
\value{

}
\description{
get_singleR_label
}
