% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC.R
\name{get_singleR_topCells}
\alias{get_singleR_topCells}
\title{get_singleR_topCells}
\usage{
get_singleR_topCells(
  singleR_out,
  topPercentileOfDeltaToKeep = 0.9,
  remove_valueLabelNonConsistent = T
)
}
\arguments{
\item{remove_valueLabelNonConsistent}{}
}
\value{

}
\description{
get_singleR_topCells
}
