% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{get_subtype_final}
\alias{get_subtype_final}
\title{get_subtype_final}
\usage{
get_subtype_final(
  id,
  df_feateure_exp = df_feateure_exp,
  df_out_phenograph = df_out_phenograph,
  df_out_svm = df_out_svm,
  out_mutation = out_mutation,
  chrom_n = chrom_n,
  CNV_label = CNV_label,
  fusion_fc = fusion_fc,
  fusion_c = fusion_c
)
}
\arguments{
\item{fusion_c}{}
}
\value{

}
\description{
get_subtype_final
}
