% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{get_vst_values}
\alias{get_vst_values}
\title{get_vst_values}
\usage{
get_vst_values(obj_in = NULL, file_obj = NULL, df_count, out_fmt = "df")
}
\arguments{
\item{obj_in}{, object, object of reference}

\item{file_obj}{, string, file of object of reference}

\item{df_count}{, dataframe, count dataframe for merge}
}
\value{

}
\description{
get_vst_values
}
