% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{get_weights}
\alias{get_weights}
\title{Weight calculation}
\usage{
get_weights(
  CNV_data,
  count_dir,
  standard_samp = NULL,
  metadata,
  CNV_data_format,
  fill_cor = TRUE
)
}
\arguments{
\item{CNV_data}{path to a table with known CNVs on arm-level. The information is used to generate CNV-gene expression level correlation coeffiecient, that are
subsequently used for gene weight calculation. The more samples are provided, the more robust the calculated correlation coeffiecient will be.
There are two optional input formats. Either in wide format, where each column represents one chromosome arm e.g. columns (1p,1q,2p,2q,3p,3q..) and
one additional column with sample name. The values in chromosome arm column are the CNV states on that chromosome arm, if the CNV state is unknown the value should be set to NA. Also the CNV_data file
can be in "long" format. Here the table is required to have three columns: sample: sample name, arm: chromosome arm (format: "1p", "19q", "Xp") and CNV: (CNV state, e.g. +1, -1, +2, 0)}

\item{standard_samp}{character vector with standard sample names, which will be used for normalization, and adjustment of diploid level. These samples should not
carry any CNVs. Providing standard samples is important for datasets, where high numbers of CNVs (e.g. 10 gained chromosomes or 10 deleted chromosomes) of one type are present is some samples.
 The sample names must be included in the first column of metadata table. If no standard samples samples are input, the data will be normalized as a batch and no diploid level adjustment will
 be performed.}

\item{metadata}{path to a metadata table with three columns. First colum: sample names, second column: file names of count files, third column: file names of snv files. There should be no header.
More information is included in the package README file.}

\item{CNV_data_format}{string, in which format is the CNV_data provided, can be either "wide" or "long"}

\item{fill_cor}{if TRUE (default), genes for which correlation coefficients cannot be estimated and and those for which adjusted p-value by Benjamini-Hochberg method is higher than 0.1, the correlation coefficient value will be set
to 0.1. This was determined as the mean gene expression-CNV correlation coefficient from qcute lymphoblastic leukemia data on all genes. This parameter may help in creating weights for datasets, for which the coverage by CNVs
of genome is low and thus correlation coeffiecients may be constructed only for small proportion of genes. If FALSE, genes without determined correlation coefficients and those for which adjusted p-value is higher than 0.1 will be excluded.}

\item{output_file}{path to a file, where gene weight matrix will be saved.}
}
\description{
Get weights for CNA analysis
}
\details{
The function provides an option to create customized weight matrix based on specific data (cancer) type. The weights are based on the correlation of gene expression with CNV and expression variance across the cohort. To calculate gene expression-CNV correlation
reliable set of CNVs on arm level is required.  For the function to work properly, each chromosome arm should have
minimum two CNV states in the provided CNV data. If this is not the case the parameter fill_cor should be set as TRUE to artificially fill out the correlation coefficients.
}
