% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_barplot}
\alias{gg_barplot}
\title{gg_barplot}
\usage{
gg_barplot(
  df_plot,
  x = NULL,
  y = NULL,
  group.by = NULL,
  strata = NULL,
  cols = cols_in_default,
  type = "identity",
  position = "stack",
  bar_width = 0.9,
  y_lab = NULL,
  x_lab = NULL,
  legend_lab = NULL,
  x_tick_label_var = NULL,
  title = ""
)
}
\arguments{
\item{title}{}
}
\value{

}
\description{
gg_barplot
}
