% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_boxPlot}
\alias{gg_boxPlot}
\title{gg_boxPlot}
\usage{
gg_boxPlot(
  df_in,
  var_value,
  var_group,
  var_strata = NULL,
  size = 0.6,
  highlightLevel = NULL,
  sizeHighlight = 1.5,
  cols = subtypeCol,
  plot_title = NULL,
  x_lab = NULL,
  y_lab = NULL
)
}
\arguments{
\item{y_lab}{}
}
\value{

}
\description{
gg_boxPlot
}
