% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_dimPlot}
\alias{gg_dimPlot}
\title{gg_dimPlot}
\usage{
gg_dimPlot(
  df_in,
  x = "uMAP_1",
  y = "uMAP_2",
  var_col = "diag",
  cols = subtypeCol(),
  size = 1,
  axis_by = 5,
  highlightLevel = NULL,
  sizeHighlight = 2,
  plot_title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  legend_title = "Group"
)
}
\arguments{
\item{legend_title}{}
}
\value{

}
\description{
gg_dimPlot
}
