% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_heatmap}
\alias{gg_heatmap}
\title{gg_heatmap}
\usage{
gg_heatmap(
  df_in,
  x = "labels",
  y = "variable",
  var_col = "value",
  y_lab = NULL,
  x_lab = NULL,
  y_tick_label_var = NULL,
  reverse_y = F,
  y_xais_side = "right",
  title = "Scores (column scaled to 0-1)"
)
}
\arguments{
\item{title}{}
}
\value{

}
\description{
gg_heatmap
}
