% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_tilePlot}
\alias{gg_tilePlot}
\title{gg_tilePlot}
\usage{
gg_tilePlot(
  df_in,
  x,
  y,
  var_col,
  cols = subtypeCol(),
  title = "Heatmap",
  y_lab = "",
  x_lab = "Number of features",
  legend_lab = "Subtype:",
  reverse_y = F,
  x_tick_label_var = NULL,
  squared = T,
  add_border = T,
  border_lwd = 0.3,
  border_linetype = 1,
  border_color = "white",
  x.axis = T
)
}
\arguments{
\item{x.axis}{}
}
\value{

}
\description{
gg_tilePlot
}
