% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilits.R
\name{obj_merge}
\alias{obj_merge}
\title{obj_merge}
\usage{
obj_merge(obj_in = NULL, file_obj = NULL, df_in, assay_name_in = "vst")
}
\arguments{
\item{obj_in}{, object, object of reference}

\item{file_obj}{, string, file of object of reference}

\item{df_in}{, dataframe, count dataframe for merge}
}
\value{

}
\description{
obj_merge
}
\examples{
obj_x=obj_merge(file_obj = "../tsne/234_featureCounts/obj_234_featureCountsMini.rds",df_in = df_counts)
}
