% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_RNAseqCNV.R
\name{plot_exp}
\alias{plot_exp}
\title{Title plot_exp}
\usage{
plot_exp(
  count_ns_final,
  box_wdt,
  sample_name,
  ylim,
  estimate,
  feat_tab_alt,
  gender
)
}
\arguments{
\item{gender}{}
}
\value{

}
\description{
Title plot_exp
}
