% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilits.R
\name{read_input}
\alias{read_input}
\title{read_input}
\usage{
read_input(file, delimiter = "\\t", header = F, count_field_num = 2)
}
\arguments{
\item{file, }{String, Input file name}

\item{delimiter, }{String, delimiter}

\item{header, }{Logical, if using the first line as header}

\item{count_field_num, }{num, num of count column in the file}
}
\value{

}
\description{
read_input
}
\examples{
read_input(file)
}
