% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{run_countMatrix}
\alias{run_countMatrix}
\title{run_countMatrix}
\usage{
run_countMatrix(file_countMatrix, featureN_PG = c(100, 1058))
}
\arguments{
\item{featureN_PG}{}
}
\value{

}
\description{
run_countMatrix
}
