% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{run_multiple_samples}
\alias{run_multiple_samples}
\title{run_multiple_samples}
\usage{
run_multiple_samples(
  file_listing,
  featureN_PG = c(100),
  minReadCnt = 3,
  minDepth = 20,
  mafmin = 0.1,
  mafmax = 0.85
)
}
\arguments{
\item{mafmax}{}
}
\value{

}
\description{
run_multiple_samples
}
