% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{run_multiple_umap}
\alias{run_multiple_umap}
\title{run_multiple_umap}
\usage{
run_multiple_umap(outs_multiple, n_neighbors = 10, variable_n = 1058)
}
\arguments{
\item{variable_n}{}
}
\value{

}
\description{
run_multiple_umap
}
