% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{run_one_sample}
\alias{run_one_sample}
\title{run_one_sample}
\usage{
run_one_sample(
  sample_id = "",
  file_count,
  file_vcf,
  file_fusioncatcher = "",
  file_cicero = "",
  featureN_PG = c(100),
  minReadCnt = 3,
  minDepth = 20,
  mafmin = 0.1,
  mafmax = 0.85
)
}
\arguments{
\item{mafmax}{}
}
\value{

}
\description{
run_one_sample
}
