% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{run_PhenoGraph}
\alias{run_PhenoGraph}
\title{run_PhenoGraph}
\usage{
run_PhenoGraph(
  obj_in,
  feature_panel = "variable_genes",
  variable_n = variable_n,
  neighbor_k = neighbor_k,
  ratio_cutoff = 0.5,
  out_label = "PhenographPred"
)
}
\arguments{
\item{out_label}{}
}
\value{

}
\description{
run_PhenoGraph
}
