% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{run_RNAseqCNV}
\alias{run_RNAseqCNV}
\title{run_RNAseqCNV}
\usage{
run_RNAseqCNV(
  df_count,
  snv_file,
  mafRange = c(0.1, 0.9),
  genome_version = "hg38",
  snv_format = "vcf",
  model_gend = model_gender,
  model_dip = model_dipl,
  model_alter = model_alt,
  model_alter_noSNV = model_noSNV,
  generate_weights = FALSE,
  minReadCnt = 3,
  samp_prop = 0.8,
  weight_samp_prop = 1,
  minDepth = 20,
  weight_tab = weight_table
)
}
\arguments{
\item{weight_tab}{}
}
\value{

}
\description{
run_RNAseqCNV
}
