% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsne.R
\name{run_tsne}
\alias{run_tsne}
\title{run_tsne}
\usage{
run_tsne(
  obj_in,
  perplexity_one = 30,
  dims = 2,
  variable_n = 800,
  out_label = "tsne",
  feature_panel = "variable_genes"
)
}
\arguments{
\item{feature_panel}{}
}
\value{

}
\description{
run_tsne
}
