% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{run_umap}
\alias{run_umap}
\title{run_umap}
\usage{
run_umap(
  obj_in,
  n_neighbors = 30,
  dims = 2,
  variable_n = 800,
  out_label = "umap",
  feature_panel = "variable_genes",
  min_dist = 0.3
)
}
\arguments{
\item{feature_panel}{}
}
\value{

}
\description{
run_umap
}
