% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{run_vst}
\alias{run_vst}
\title{run_vst}
\usage{
run_vst(
  obj_in,
  assay_name_in = "counts",
  assay_name_out = "vst",
  var_design = NULL
)
}
\arguments{
\item{obj_in}{, object, object of reference}

\item{assay_name_in}{, string, assay name used for running vst}

\item{assay_name_out}{, string, assay names used for storing the vst values}

\item{var_design}{, string, variable in coldata used in vst running formula design. will not affecting the vst values}
}
\value{

}
\description{
run_vst
}
\examples{
obj_x=run_vst(obj_in = obj_x)
}
