% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_server.R
\name{server_singleSample}
\alias{server_singleSample}
\title{server_singleSample}
\usage{
server_singleSample(input, output, session)
}
\arguments{
\item{session}{}
}
\value{

}
\description{
server_singleSample
}
