% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{get_RNAseqCNV_plot}
\alias{get_RNAseqCNV_plot}
\title{Title get_RNAseqCNV_plot}
\usage{
get_RNAseqCNV_plot(
  RNAseqCNV_out,
  sample_name = "TestSample",
  adjust = TRUE,
  estimate_lab = TRUE,
  scale_cols = scaleCols,
  dpRatioChromEdge = dpRatioChrEdge,
  size1 = 3,
  size2 = 3
)
}
\arguments{
\item{scale_cols}{}
}
\value{

}
\description{
Title get_RNAseqCNV_plot
}
