% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{plot_exp}
\alias{plot_exp}
\title{Title plot_exp}
\usage{
plot_exp(
  count_ns_final,
  box_wdt,
  sample_name,
  ylim,
  estimate,
  feat_tab_alt,
  gender,
  size1 = 3,
  size2 = 3
)
}
\arguments{
\item{gender}{}
}
\value{

}
\description{
Title plot_exp
}
